Refactoring time_series and added some unit tests
